+++ decidim-bulletin_board-ruby/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## Added
+
+- The `Voter` returns the ballot hash after encrpyting the plain vote and before auditing or casting it
+
 ## [0.11.0] - 2021-02-18
 
 ## Changed